Add Orleans.Runtime.Abstractions projects Move interfaces and types needed for extensions from Orleans.Runtime to Orleans.Runtime.Abstractions.
