+++ Orleans.sln
@@ -182,6 +182,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OrleansSQLUtils", "src\AdoN
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AdoNet", "AdoNet", "{EB2EDE59-5021-42EE-A97A-D59939B39C66}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Runtime.Abstractions", "src\Orleans.Runtime.Abstractions\Orleans.Runtime.Abstractions.csproj", "{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1116,7 +1118,6 @@ Global
 		{3E2C5440-D6C5-4EB8-8D34-10707B322263}.Release|x64.Build.0 = Release|Any CPU
 		{3E2C5440-D6C5-4EB8-8D34-10707B322263}.Release|x86.ActiveCfg = Release|Any CPU
 		{3E2C5440-D6C5-4EB8-8D34-10707B322263}.Release|x86.Build.0 = Release|Any CPU
-
 		{57004849-FB72-4278-B587-671904ED455C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{57004849-FB72-4278-B587-671904ED455C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57004849-FB72-4278-B587-671904ED455C}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1129,7 +1130,6 @@ Global
 		{57004849-FB72-4278-B587-671904ED455C}.Release|x64.Build.0 = Release|Any CPU
 		{57004849-FB72-4278-B587-671904ED455C}.Release|x86.ActiveCfg = Release|Any CPU
 		{57004849-FB72-4278-B587-671904ED455C}.Release|x86.Build.0 = Release|Any CPU
-
 		{14E79ED0-4DFB-43B5-879B-1DD6A5A96473}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{14E79ED0-4DFB-43B5-879B-1DD6A5A96473}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{14E79ED0-4DFB-43B5-879B-1DD6A5A96473}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1154,6 +1154,18 @@ Global
 		{3CBF459A-C43F-40B7-8F96-2B44D55840C5}.Release|x64.Build.0 = Release|Any CPU
 		{3CBF459A-C43F-40B7-8F96-2B44D55840C5}.Release|x86.ActiveCfg = Release|Any CPU
 		{3CBF459A-C43F-40B7-8F96-2B44D55840C5}.Release|x86.Build.0 = Release|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Debug|x64.Build.0 = Debug|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Debug|x86.Build.0 = Debug|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Release|x64.ActiveCfg = Release|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Release|x64.Build.0 = Release|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Release|x86.ActiveCfg = Release|Any CPU
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1246,6 +1258,7 @@ Global
 		{14E79ED0-4DFB-43B5-879B-1DD6A5A96473} = {EB2EDE59-5021-42EE-A97A-D59939B39C66}
 		{3CBF459A-C43F-40B7-8F96-2B44D55840C5} = {EB2EDE59-5021-42EE-A97A-D59939B39C66}
 		{EB2EDE59-5021-42EE-A97A-D59939B39C66} = {FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A}
+		{9A3CB1AD-1B58-4D89-823D-9CBD00679F77} = {4CD3AA9E-D937-48CA-BB6C-158E12257D23}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7BFB3429-B5BB-4DB1-95B4-67D77A864952}
