+++ src/Orleans.Runtime/Hosting/Generic/HostingEnvironment.cs
@@ -11,28 +11,6 @@ namespace Orleans.Hosting
     }
 
     /// <summary>
-    /// Constants for HostBuilder configuration keys.
-    /// </summary>
-    public static class HostDefaults
-    {
-        /// <summary>
-        /// The configuration key used to set <see cref="IHostingEnvironment.ApplicationName"/>.
-        /// </summary>
-        public static readonly string ApplicationKey = "applicationName";
-
-        /// <summary>
-        /// The configuration key used to set <see cref="IHostingEnvironment.EnvironmentName"/>.
-        /// </summary>
-        public static readonly string EnvironmentKey = "environment";
-
-        /// <summary>
-        /// The configuration key used to set <see cref="IHostingEnvironment.ContentRootPath"/>
-        /// and <see cref="IHostingEnvironment.ContentRootFileProvider"/>.
-        /// </summary>
-        public static readonly string ContentRootKey = "contentRoot";
-    }
-
-    /// <summary>
     /// Commonly used environment names.
     /// </summary>
     public static class EnvironmentName
