+++ src/Orleans.Runtime/MultiClusterNetwork/MultiClusterOracle.cs
@@ -454,7 +454,7 @@ namespace Orleans.Runtime.MultiClusterNetwork
             return result;
         }
 
-        private void PublishMyStatusToNewDestinations(MultiClusterData delta)
+        private void PublishMyStatusToNewDestinations(IMultiClusterGossipData delta)
         {
             // for quicker convergence, we publish active local status information
             // immediately when we learn about a new destination
@@ -529,7 +529,7 @@ namespace Orleans.Runtime.MultiClusterNetwork
             // set this flag to request a full gossip (synchronize)
             protected bool doSynchronize = false;
 
-            public void Publish(MultiClusterData data)
+            public void Publish(IMultiClusterGossipData data)
             {
                 // add the data to the data waiting to be published
                 toPublish = toPublish.Merge(data);
